feat(atom::store::QueryStore): implement gix::Url
The `gix::Url` implementation of the `QueryStore` is markedly more
efficient than other implements, as it does not require a local
repository and does not fetch any git objects.

It leverages low-level transport and handshake primitives to make the
operation as efficient as possible to enable high speed querying of
many repositories efficiently.