fix(lock): finish lock synchronization for atoms
For those atom's which are not being removed, and are not being
requested, we just need to make a pass that their VersionReq in the
manifest has not fallen out of alignment with the lock.

If it has, in order for the lock to remain always consistent, we need
to re-resolve them.

For those atoms in the manifest but not the lock, we also need to ensure
they are resolved to ensure the lock and manifest are always
syncrhonized.

This change introduces this logic, when coupled with the sanitization
logic of previous changesets, this essentially completes manifest/lock
syncrhonization on each call to `eka add`.