chore: prepare to merge
Make some critical refinements to the public API, mainly that
dependencies on the manifest and lockfile struct are not public to the
outside world so that modifications can only be made through the proper
interfaces which ensure they remain in sync.

Also refine the documentation up to the crate standard, and focus on
updating ADR #2 with details from the actual implementation which has
now diverged significantly from the original plan.