fix(uri): fix aliased url parsing
We want to support the more familiar https://foo.com/bar@master syntax
to specify git references, but it requires special handling to avoid
overlapping with url usernames.

To solve this, parse the url with gix::url first, which can distinguish
a username from a `@` in a url path, then parse to a UriRef so we can
also expand aliases.

Also fix distinguishing atom uris from urls more directly by simply
searching for an atom's `::` delimiter.