style: major style overhaul
In preparation of outside contributors, align the entire codebase with
our new STYLE_GUIDE.md. Also add a CONTRIBUTING.md to give additional
information and context on how to contribute, as well as an ISSUE and
PR template to get us started.

Finally, a simple ADR template was also included to give anyone who
might like to propose something a simple starting point.