feat: implement git pin synchronization
If a git pin has been changed in a way where it's locked url no longer
matches what's in the manifest, or if an explicit version was requested
in the manifest and no longer matches what's locked: re-resolve.

This explicitly skips the case where a ref in the manifest points to
a locked rev, as this should only be updated if explicitly requested by
the user (e.g. future `eka resolve update`).