refactor(sets): restructure set detection logic
The previous structure was extremely ad-hoc and difficult to follow.

Introduce a new `SetResolver` struct for managing the context and reduce
function arguments. This helps the overall readability of mirror
resolution logic, which is critical for maintenance.

Also wires up the results of this resolution with the `ManifestWriter`,
in preparation for leveraging these results to refactor the lock
sychronization code to conform to the new schema outline in ADR#7.