chore(lock): prep for synchronization code
Restructure collated atom responses from mirrors to make lock
synchronization code more reasonable.

Also moves the sync and sanitization code directly onto the
ManifestWriter struct, as it's more appropriate.

Also modifies the hash type in the id module in order to avoid a
duplicate type in the lock, allowing us to more naturally leverage the
`AtomId` type as a HashMap key, and as a result from the `QueryVersion`
trait.