Revert "Merge pull request #20 from ekala-project/publish-init"
This reverts commit e07378439d5e012d3cf77176c588a3a97677cf4c, reversing
changes made to 051f0d893f08e44fc183805d6619c64033b7afbe.

This is due to the fact that with the new `ekala.toml`, the `init`
subcommand becomes a bit more than a simple atom publishing concern, as
it now serves to initialize the canonical "index" location for declaring
atoms in this set.