docs(atom): Update tag validation docs to reflect UAX #31
The `id` module was recently updated to use the Unicode UAX #31
standard for identifier validation, via the `unicode_ident` crate. This
change also introduced unconditional NFKC normalization.

This commit updates the documentation comments to accurately reflect this
new validation logic, making it clear that we are conforming to a
specific standard. It also clarifies the single exception to the
standard: allowing the `-` character in the body of an identifier.