fix(resolve): fixes for local path resolution
local dependencies were being remove from the lock inappripriately on
account of not calling `return` in the retain method and falling through
to false.

Also, we need to ensure in the first case of atom_uri that we are
actually working with a remote, so that local paths don't enter this
branch and fail to resolve.

Also factors out the local resolution lock to be reused in lock sync.