fix(add): fix the deserializer for nix dependencies
NixFetch was failing to deserialize due to the attribute
`deny_unknown_keys`, on a flattened key which itself was represented
by a complex untagged enum.

Since the field name varied by variant, it failed uncondtionally.
Simply removing the attribute solved the issue and deserialization is
now working for this type.