fix(ekala): properly report manifest load failure
Previously we were just throwing a generic "no such file" error when
we tried to open the repositories ekala.toml, which would have been
very confusing to the user without additional context. Now we report
specifically that it is the manifest itself that we have failed to
locate and give a helpful suggestion.