feat(lock): make lock synchronization a hard error
Though this constitutes a very small code change, it indicates a
significant milestone in development.

Lock synchronization has reached a significantly robust state, and though
it is possible bugs still exist, it is complete enough to enforce and
fail early to encourage users to keep their manifests coherent.

Also, if there are any outstanding bugs, this will force us to become
aware of them sooner than later.

The specific synchronization error will be printed along with the
relevant details about the specific atom which failed.