fix: decouple atom commit from incidental path
The path information is totally agnostic to atom. In the same way atoms
are "resolved" from a remote, they are now "resolved" locally from the
ekala.toml.

This means an atom can be relocated in the repository at will without
modifying it's content hash, so long as nothing else changes.