refactor(treewide): reorganize code in more modules
Some of the modules were becoming very large, so we begin to refactor a
large portion of code in various modules to hopefully more sensibly
begin to organize the logic into intuitive categories.

More work is needed to clean up the modules in their new hierachies, and
many of the docs tests are now broken. Follow up commits will address
these outstand concerns.