feat(manifest): write singleton k/v type
Declares a `Singleton` type for expressing an arbitrary key value pair
of precisely one unit. This allows for serializing and deserializing
transparently as a Map while enforcing the "singleton" aspect at the
type level.

Also declare a `Composer` and `ComposerSet` as concrete instances of
this generic `Singleton` for use in the forthcoming composer API.

This will ensure that the manifest always declares one, and only one,
atom composer; leaving no room for ambiguity.