feat: add procedural macro to compute eka's root
The lock expression, unlike the composer, is tightly coupled to a
specific eka version. It is critical that we do not have any implicit
dependencies that could effect the derivation hash without being
specified, so the lock expression needs to, itself, be locked in the
lockfile (for nix atoms that evaluate at least).

In order to do this, while remaining tightly coupled to our specific
eka build, we will have to compute eka's own repository root hash at
compile time, and use it to expect a lock expression atom to come from
this set.

Users will be able to set mirrors for eka in their configuration files,
so long as the final lock expression atom matches what is expected.

This change begins the work to accomplish this by creating a procedural
macro which computes eka's root hash and sets it to a constant in the
atom crate.