fix(manifest): writing sets was broken
after the last fix, set writing to the manifest was broken inadvertently
by taking ownership of the mutable reference, thereby disconnecting
further changes from the underlying manifest.

Instead, set the value of the mutable reference to our new table in
the special case where there is only one value, thereby maintaining
a reference to write to.

Also replaces tabs with spaces in the ekala.toml writer, and fixes a
deserialization error when sets are empty in the lockfile.