fix(init): don't fail before init local storage
We were erroring out on storage detection because there was no None
variant for the detection enum, but we need to be able to distinguish
this in the case where we want to initialize a new local storage
directory (non-git).

To fix this simply add a None variant to the detection enum so that we
can properly match on it during initialization, that way we don't error
out too early before we can get there.