feat(nixec): setup restricted evaluation
Inside our sandbox env this works very similarly to pure evaluation
without the "copy the world" requirement.

Basically, any path outside of the current directory is invisible to
the evaluation and eval will abort with an error if you attempt to
access them.

The only remaining "holes" are impure builtins of nix expressions, such
as `builtins.getEnv` which are plugged by atom-nix creating a totally
pure evaluation context for much less cost.

Currently this change hardcodes allowed uris, since they are required
to specify when setting restricted eval. In the future we should allow
setting these externally, either via a flag or perhaps an environmental
variable.