release(nix-lock): 0.1.6 -> 0.2.0
Adds a new "import.nix" expression used to import this specific version
of the larger lock expressions. The import.nix will be read at eka
compile time and stored in the binary itself, making it convenient to
call into atoms without any implicit nix code, and without making the
baked in expression too rigid, since it is versions and well specified
in the lock.

The fact that we bake this particular expression into the lock can be
considered an "implementation detail" to make calling it more straight
forward.