fix(git): retry if transports can't persist
Some transports cannot persist multiple connections. We want to be able
to maximally reuse transports whenever we can while still degrading
gracefully when we have a transport implement which cannot.

So during fetching, we call the network as usual, but if it fails, we
check if the original transport we were handed is not able to persist
via the `connection_persists_across_multiple_requests` method, and if
not, we try again with a fresh transport.

This avoids pointless retries if the connection failed for some other
reason while allowing us to handle those transports in gix which do not
currently support more than one request. Since we cannot know if our
passed transport has already been used beforehand, we attempt the network
call to find out first, then check.