fix(git): get_ref needs to be filtered
Even though git only fetches the objects specified in the target, the
server can actually return more refs than requested as an optimization.

For this reason, the call to `next()` would return the wrong ref.
Instead we use `find` to ensure the ref we return matches the one we
actually requested from the server.