feat(plan): make remote work asychronous
In order that we can setup atoms as fast as possible, make the remote
calls in a JoinSet, and since gix transports are still blocking, use a
blocking thread pool to allow new asyc tasks to continue before
completion.

Also, remove writing the remote to the cache repositories configuration.
It doesn't really gain us anything right now, and causes a race
condition.