feat!: change atom lock schema
removes the locker entry entirely in place of a more abstract API. From
now on, eka will use the version of nix-lock which corresponds to the
format version as a semver major version.

In other words, format version 1 should work with any version 1.* of the
nix-lock atom. Since we are still unstable, the format version has, for
now, been changed to 0 to indicate that we may change at any time.

This change also makes the composer act more like a regular dependency
(it's placed in the deps list), and only a symbolic reference to that
underlying dependency is placed in the lock via the atom id and version
to signify to the locking expression which atom to treat specially as
our composer.