feat: remove `mirror` field from lock
We will expect nix to resolve atoms from the atom store instead of
directly from source. Since we will have the full transitive closure
in the lockfile after we integrate SAT resolution, we no longer need
this field, and it was always meant as a stop gap.

In preparation for integrating the full scale resolver, we remove it now.